requirements.txt
bottle==0.12.7
psutil==2.1.1
python-gnupg==0.3.6
requests==2.2.1
scrypt==0.6.1
tornado==3.2.2
tempdir==0.6
docker-py==0.3.2
